CP-204 Sorting functionality in Tender table
- [x] Tender table; only in selected columns
