+++ package.json
@@ -74,6 +74,7 @@
   },
   "devDependencies": {
     "eslint": "^7.12.1",
+    "redux-devtools-extension": "^2.13.8",
     "eslint-config-standard": "^16.0.1",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-node": "^11.1.0",
